+++ main.py
@@ -113,6 +113,7 @@ set_money_activation = True
 get_all_data_activation = True
 reset_user_account_activation = True
 coinflip_activation = True
+reward_activation = True
 
 #Tous les status que vas prendre le bot
 status = cycle([
@@ -165,7 +166,7 @@ bot.add_cog(moderation.Moderation(bot,kick_activation,delete_activation,prefix))
 
 bot.add_cog(autres.Autres(bot,prefix,commands_activation))
 
-bot.add_cog(economy.Economy(bot,eco,prefix,bal_activation, work_activation,pay_activation,bet_activation,daily_activation,add_money_activation,remove_money_activation,set_money_activation,get_all_data_activation,reset_user_account_activation,coinflip_activation))
+bot.add_cog(economy.Economy(bot,eco,prefix,bal_activation, work_activation,pay_activation,bet_activation,daily_activation,add_money_activation,remove_money_activation,set_money_activation,get_all_data_activation,reset_user_account_activation,coinflip_activation, reward_activation))
 
 bot.add_cog(zone_de_test.Zone_de_test(bot,prefix, dice_activation))